requirements.txt
uWSGI==2.0.17
aniso8601==3.0.0
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
class-registry==2.1.2
click==6.7
configobj==5.0.6
cryptography==2.2.2
filters==1.3.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
flask-mongoengine==0.9.5
flask-restplus==0.10.1
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mongoengine==0.15.0
pycparser==2.18
PyJWT==1.6.1
pymongo==3.6.1
pyOpenSSL==17.5.0
PyOTA==2.0.5
pysha3==1.0.2
python-dateutil==2.7.2
pytz==2018.3
regex==2018.2.21
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.47.0
Werkzeug==0.14.1
WTForms==2.1
